Crate objc2_core_video

Bindings to the CoreVideo framework

See Apple’s docs and the general docs on framework crates for more information.

Structs

CVSMPTETime

A structure for holding a SMPTE time. Field: subframes The number of subframes in the full message. Field: subframeDivisor The number of subframes per frame (typically 80). Field: counter The total number of messages received. Field: type The kind of SMPTE time using the SMPTE time type constants. Field: flags A set of flags that indicate the SMPTE state. Field: hours The number of hours in the full message. Field: minutes The number of minutes in the full message. Field: seconds The number of seconds in the full message. Field: frames The number of frames in the full message.

CVSMPTETimeFlags

Flags that describe the SMPTE time state.

CVSMPTETimeType

Constants that describe the type of SMPTE time.

CVTime

Apple’s documentation

CVTimeFlags

Apple’s documentation

CVTimeStamp

CoreVideo uses a CVTimeStamp structure to store video display time stamps.

CVTimeStampFlags

Apple’s documentation

Statics

kCVIndefiniteTime⚠

Apple’s documentation

kCVZeroTime⚠

Apple’s documentation

Type Aliases

CVOptionFlags

Flags to be used for the display and render call back functions.